backend/users/views/authentication.py
from django.contrib.auth import logout
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response


@api_view(["GET"])
@permission_classes([IsAuthenticated])
def logout_view(request, *args, **kwargs):
    logout(request)
    return Response(None, 204)
