backend/pyproject.toml
[tool.isort]
profile = "black"

[tool.black]
line-length = 130

[tool.mypy]
mypy_path = "src/"

[tool.pytest.ini_options]
env = [
    "DJANGO_SECRET_KEY=secret",
    "RECAPTCHA_SECRET=secret"
]
DJANGO_SETTINGS_MODULE = "lionskins.settings"

[tool.poetry]
name = "lionskins"
version = "1.0.0"
description = ""
authors = ["Julien <contact@lionskins.co>"]

[tool.poetry.dependencies]
python                        = "^3.10"
beautifulsoup4                = "~4.10.0"
click                         = "~8.0.1"
django                        = "~4.0.0"
djangorestframework           = "~3.13.1"
django-apscheduler            = "~0.6.2"
django-cors-headers           = "~3.11.0"
django-filter                 = "~21.1"
graphene-django               = "~3.0.0b7"
psycopg2-binary               = "~2.9.3"
pymongo                       = "*"
python-dotenv                 = "~0.19"
python-slugify                = "~6.1.1"
ratelimit                     = "~2.2.1"
requests                      = "~2.27"
social-auth-app-django        = "~5.0.0"
structlog                     = "~21.5"
tqdm                          = "~4.63"
user-agents                   = "~2.2.0"

# backoffice
apscheduler                 = "~3.9.1"

# bitskins
pyotp                       = "~2.6.0"

# deployment
gunicorn                    = "~20.1.0"

# sentry
sentry-sdk                  = {version = "~1.5", extras = ["django"]}

[tool.poetry.dev-dependencies]
mypy                        = "*"
pre-commit                  = "*"
factory_boy                 = "~3.2"
pytest                      = "~7.1"
pytest-cov                  = "~3.0"
pytest-env                  = "~0.6"
pytest-factoryboy           = "~2.1"
pytest-django               = "~4.5"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
